third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/forms/the-option-element/option-selected-expected.txt

This is a testharness.js-based test.
PASS not dirty 
FAIL dirty, selected assert_equals: expected true but got false
FAIL dirty, not selected assert_equals: expected false but got true
Harness: the test ran to completion.
